Remove keyboard event listener, `quoteSelection` method.
BREAKING CHANGE

The code for handling eventlistening is very trivial, and might vary per
consumer of this library. It is not core functionality, and should be
left to upstream to setup.

Removing this reduces the API surface area and complexity from this library
as it doesn't need to concern itself with event management and keyboard
shortcuts.

Co-authored-by: Keith Cirkel <keithamus@users.noreply.github.com>