Remove `getSelectionContext()` argument, add `quoteElement`.
We can remove `getSelectionContext` entirely, by inlining it into the
`extractQuote` body, however it was useful to pass in a custom selection
context for the use case of selecting the contents of an element.

By introducing `quoteElement` as an option we can supply the single use
case for creating custom `SelectionContext`s, which also can replace
some code we have upstream.